# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  entity.UserCreate:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    type: object
  entity.UserUpdate:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      role:
        type: string
    type: object
  entity.WebsiteCreate:
    properties:
      description:
        type: string
      domain:
        type: string
      logo_url:
        type: string
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
  entity.WebsiteUpdate:
    properties:
      description:
        type: string
      domain:
        type: string
      is_active:
        type: boolean
      logo_url:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@news-api.dev
    name: News API Support
  description: REST API untuk website berita. Mendukung auth JWT, artikel, kategori,
    tag, dan komentar.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: News API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login dengan email & password, mengembalikan JWT token
      parameters:
      - description: Kredensial login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login sukses, berisi token & user
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Email atau password salah
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 'Mendaftarkan user baru (default role: author)'
      parameters:
      - description: Data registrasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: User berhasil dibuat
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Register user baru
      tags:
      - Auth
...
  /uploads/images:
    post:
      consumes:
      - multipart/form-data
      description: Upload gambar (jpeg/png/webp/gif/avif, max 5MB). Mengembalikan
        URL publik untuk dipakai di artikel.
      parameters:
      - description: File gambar (jpeg/png/webp/gif/avif)
        in: formData
        name: file
        required: true
        type: file
      - description: 'Folder tujuan (default: uploads)'
        in: formData
        name: folder
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Upload berhasil, berisi url & key
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: File tidak valid
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Upload gambar
      tags:
      - Uploads
  /users:
    get:
      description: Menampilkan daftar user (hanya admin)
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Daftar user
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Menghapus user (hanya admin)
      parameters:
      - description: ID user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User berhasil dihapus
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Mengubah data user (hanya admin)
      parameters:
      - description: ID user
        in: path
        name: id
        required: true
        type: string
      - description: Data yang diubah
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: User berhasil diubah
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /users/me:
    get:
      description: Mengambil data user berdasarkan token JWT
      produces:
      - application/json
      responses:
        "200":
          description: Data user
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Tidak terautentikasi
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Profil user saat ini
      tags:
      - Users
  /websites:
    get:
      description: Menampilkan daftar semua website dengan pagination
      parameters:
      - default: 1
        description: Nomor halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Daftar website dengan meta pagination
          schema:
            $ref: '#/definitions/response.Response'
      summary: List website
      tags:
      - Websites
    post:
      consumes:
      - application/json
      description: 'Membuat website baru (role: admin). Setiap website punya artikel
        sendiri-sendiri.'
      parameters:
      - description: Data website
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.WebsiteCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Website berhasil dibuat
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Request tidak valid
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat website
      tags:
      - Websites
  /websites/{id}:
    delete:
      description: 'Menghapus website beserta seluruh artikelnya (role: admin)'
      parameters:
      - description: ID website
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Website berhasil dihapus
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Website tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus website
      tags:
      - Websites
    get:
      description: Mengambil detail website berdasarkan ID
      parameters:
      - description: ID website
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data website
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Website tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail website by ID
      tags:
      - Websites
    put:
      consumes:
      - application/json
      description: 'Mengubah data website (role: admin)'
      parameters:
      - description: ID website
        in: path
        name: id
        required: true
        type: string
      - description: Data yang diubah
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.WebsiteUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Website berhasil diubah
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Website tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update website
      tags:
      - Websites
  /websites/slug/{slug}:
    get:
      description: Mengambil detail website berdasarkan slug
      parameters:
      - description: Slug website
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data website
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Website tidak ditemukan
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail website by slug
      tags:
      - Websites
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"